refactor : adapt to directory structure change and code refinement procedure
