+++ src/main/java/com/softgallery/issuemanagementbackEnd/controller/chat_gpt/ChatGptController.java
@@ -1,11 +1,7 @@
-package com.softgallery.issuemanagementbackEnd.controller.chatGpt;
+package com.softgallery.issuemanagementbackEnd.controller.chat_gpt;
 
-import com.softgallery.issuemanagementbackEnd.dto.ChatGptRequestDTO;
-import com.softgallery.issuemanagementbackEnd.dto.ChatGptResponseDTO;
-import com.softgallery.issuemanagementbackEnd.dto.IssueDTO;
-import com.softgallery.issuemanagementbackEnd.dto.QuestionRequestDTO;
+import com.softgallery.issuemanagementbackEnd.dto.chat_gpt.ChatGptResponseDTO;
 import com.softgallery.issuemanagementbackEnd.service.chatGpt.ChatGptServiceIF;
-import org.springframework.http.HttpEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
