+++ src/main/java/com/softgallery/issuemanagementbackEnd/dto/chat_gpt/QuestionRequestDTO.java
@@ -1,4 +1,4 @@
-package com.softgallery.issuemanagementbackEnd.dto;
+package com.softgallery.issuemanagementbackEnd.dto.chat_gpt;
 
 import com.softgallery.issuemanagementbackEnd.service.chatGpt.GptPrompt;
 import lombok.Getter;
@@ -15,7 +15,7 @@ public class QuestionRequestDTO implements Serializable {
     private final double temperature=1.0;
     private final double topP=1.0;
 
-    public QuestionRequestDTO(){}
+    public QuestionRequestDTO(){ }
     public QuestionRequestDTO(String s){
         this.messages=new ArrayList<Message>();
         Message prompMessage = new Message("system",GptPrompt.getPromptVer1());
