+++ src/main/java/com/softgallery/issuemanagementbackEnd/service/chatGpt/ChatGptServiceIF.java
@@ -1,8 +1,7 @@
 package com.softgallery.issuemanagementbackEnd.service.chatGpt;
 
-import com.softgallery.issuemanagementbackEnd.dto.ChatGptRequestDTO;
-import com.softgallery.issuemanagementbackEnd.dto.ChatGptResponseDTO;
-import com.softgallery.issuemanagementbackEnd.dto.QuestionRequestDTO;
+import com.softgallery.issuemanagementbackEnd.dto.chat_gpt.ChatGptRequestDTO;
+import com.softgallery.issuemanagementbackEnd.dto.chat_gpt.ChatGptResponseDTO;
 import org.springframework.http.HttpEntity;
 
 public interface ChatGptServiceIF {
